rentup_backend/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
autopep8==2.0.0
braintree==4.15.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.11.0
django-dia==0.4.1
django-extensions==3.1.5
django-oauth-toolkit==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.4
graphviz==0.20
h11==0.13.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
oauthlib==3.2.2
outcome==1.1.0
packaging==21.3
Pillow==8.4.0
pycodestyle==2.9.1
pycparser==2.21
pydot==1.4.2
pydotplus==2.0.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
wsproto==1.1.0
